// src/__generated__/findChildCareProviders.ts
export interface findChildCareProviders_findChildCareProviders_ProviderSearchSuccess_providers_logo {
  __typename: "ProfileImage";
  /**
   * Represent original size of image, expected to be > 1080px
   */
  urlOriginal: globalScalarURL | null;
}

export interface findChildCareProviders_findChildCareProviders_ProviderSearchSuccess_providers_images {
  __typename: "ProfileImage";
  /**
   * Represent original size of image, expected to be > 1080px
   */
  urlOriginal: globalScalarURL | null;
}

export interface findChildCareProviders_findChildCareProviders_ProviderSearchSuccess_providers_address {
  __typename: "Address";
  /**
   * Address line one containing house number
   */
  addressLine1: string;
  /**
   * Address line two containing street address
   */
  addressLine2: string | null;
  /**
   * City of the address.
   */
  city: string;
  /**
   * Latitude
   */
  latitude: number | null;
  /**
   * Longitude
   */
  longitude: number | null;
  /**
   * State for the address.
   */
  state: string;
  /**
   * Zip of the address.
   */
  zip: string;
}

export interface findChildCareProviders_findChildCareProviders_ProviderSearchSuccess_providers_reviews_reviewer {
  __typename: "Reviewer";
  /**
   * Given/first name of review creator
   */
  firstName: string | null;
  /**
   * Family/last name of review creator
   */
  lastName: string | null;
}

export interface findChildCareProviders_findChildCareProviders_ProviderSearchSuccess_providers_reviews {
  __typename: "Review";
  /**
   * Review comment, free text provided by reviewer
   */
  text: string | null;
  /**
   * Date when review was submitted
   */
  date: globalScalarLocalDate | null;
  /**
   * Rating from reviewer
   */
  rating: number | null;
  /**
   * Reviewer information
   */
  reviewer: findChildCareProviders_findChildCareProviders_ProviderSearchSuccess_providers_reviews_reviewer | null;
}

export interface findChildCareProviders_findChildCareProviders_ProviderSearchSuccess_providers_license {
  __typename: "License";
  /**
   * License human readable (display) name
   */
  name: string | null;
  /**
   * Date of last verification of license
   */
  verifiedDate: globalScalarLocalDate | null;
  /**
   * Indicate if  Provider certified to provide required service
   */
  certified: boolean | null;
  /**
   * License website url
   */
  externalUrl: string | null;
  /**
   * License administrative area, state name for US
   */
  administrativeArea: string | null;
}

export interface findChildCareProviders_findChildCareProviders_ProviderSearchSuccess_providers {
  __typename: "Provider";
  /**
   * Provider id
   */
  id: string;
  /**
   * Description of provider to be showed to Seeker
   */
  description: string | null;
  /**
   * Provider name
   */
  name: string | null;
  /**
   * Different types of centers
   */
  centerType: CenterType | null;
  /**
   * Provider's logo image
   */
  logo: findChildCareProviders_findChildCareProviders_ProviderSearchSuccess_providers_logo | null;
  /**
   * Provider images like photo, logo, etc
   */
  images: findChildCareProviders_findChildCareProviders_ProviderSearchSuccess_providers_images[] | null;
  /**
   * Provider address
   */
  address: findChildCareProviders_findChildCareProviders_ProviderSearchSuccess_providers_address | null;
  /**
   * Provider reviews
   */
  reviews: (findChildCareProviders_findChildCareProviders_ProviderSearchSuccess_providers_reviews | null)[] | null;
  /**
   * Provider license, can be optional for individual providers
   */
  license: findChildCareProviders_findChildCareProviders_ProviderSearchSuccess_providers_license | null;
}

export interface findChildCareProviders_findChildCareProviders_ProviderSearchSuccess {
  __typename: "ProviderSearchSuccess";
  /**
   * ID generated by search engine to connect search request with future manipulations with Providers in result
   */
  trackingId: string;
  /**
   * List of provider with details in required area
   */
  providers: (findChildCareProviders_findChildCareProviders_ProviderSearchSuccess_providers | null)[] | null;
}

export interface findChildCareProviders_findChildCareProviders_ProviderSearchError {
  __typename: "ProviderSearchError";
  /**
   * Summary of the error.
   */
  message: string | null;
}

export type findChildCareProviders_findChildCareProviders = findChildCareProviders_findChildCareProviders_ProviderSearchSuccess | findChildCareProviders_findChildCareProviders_ProviderSearchError;

export interface findChildCareProviders {
  /**
   * Get list of Day Care Providers in specific area by zipcode and children DoBs
   */
  findChildCareProviders: findChildCareProviders_findChildCareProviders;
}

export interface findChildCareProvidersVariables {
  zipcode: string;
  childrenDoB: globalScalarLocalDate[];
  source: string;
  careStartDate?: globalScalarLocalDate | null;
  distanceWillingToTravel?: DistanceInput | null;
}